refactor for crates publishing
